requirements.txt
pandas==2.2.2
torch==2.3.1
torchtext==0.18.0
hydra-core==1.3.2
pytorch-lightning==2.2.5
numpy==1.26.4
